refactor(exceptions): organiza exceções e validações em pacote dedicado
