+++ src/main/java/com/project/miinhareceita/favorite/service/FavoriteService.java
@@ -7,9 +7,9 @@ import com.project.miinhareceita.favorite.dto.FavoriteInsertDTO;
 import com.project.miinhareceita.favorite.repository.FavoriteRepository;
 import com.project.miinhareceita.recipe.domain.Recipe;
 import com.project.miinhareceita.recipe.repository.RecipeRepository;
-import com.project.miinhareceita.services.exceptions.DatabaseException;
-import com.project.miinhareceita.services.exceptions.ForbiddenException;
-import com.project.miinhareceita.services.exceptions.ResourceNotFoundException;
+import com.project.miinhareceita.shared.exceptions.DatabaseException;
+import com.project.miinhareceita.shared.exceptions.ForbiddenException;
+import com.project.miinhareceita.shared.exceptions.ResourceNotFoundException;
 import com.project.miinhareceita.user.domain.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
