+++ src/main/java/com/project/miinhareceita/ingredient/service/IngredientService.java
@@ -4,8 +4,8 @@ import com.project.miinhareceita.ingredient.domain.Ingredients;
 import com.project.miinhareceita.ingredient.dto.IngredientDTO;
 import com.project.miinhareceita.ingredient.projection.IngredientProjection;
 import com.project.miinhareceita.ingredient.repository.IngredientsRepository;
-import com.project.miinhareceita.services.exceptions.DatabaseException;
-import com.project.miinhareceita.services.exceptions.ResourceNotFoundException;
+import com.project.miinhareceita.shared.exceptions.DatabaseException;
+import com.project.miinhareceita.shared.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
