+++ src/main/java/com/project/miinhareceita/review/service/ReviewService.java
@@ -7,9 +7,9 @@ import com.project.miinhareceita.review.domain.Review;
 import com.project.miinhareceita.review.dto.ReviewDTO;
 import com.project.miinhareceita.review.projection.ReviewProjections;
 import com.project.miinhareceita.review.repository.ReviewRepository;
-import com.project.miinhareceita.services.exceptions.DatabaseException;
-import com.project.miinhareceita.services.exceptions.ForbiddenException;
-import com.project.miinhareceita.services.exceptions.ResourceNotFoundException;
+import com.project.miinhareceita.shared.exceptions.DatabaseException;
+import com.project.miinhareceita.shared.exceptions.ForbiddenException;
+import com.project.miinhareceita.shared.exceptions.ResourceNotFoundException;
 import com.project.miinhareceita.user.domain.User;
 import com.project.miinhareceita.user.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
