+++ src/main/java/com/project/miinhareceita/shared/exceptions/FieldMessage.java
@@ -1,4 +1,4 @@
-package com.project.miinhareceita.controllers.exceptions;
+package com.project.miinhareceita.shared.exceptions;
 
 import java.io.Serializable;
