+++ src/main/java/com/project/miinhareceita/shared/exceptions/ForbiddenException.java
@@ -1,4 +1,4 @@
-package com.project.miinhareceita.services.exceptions;
+package com.project.miinhareceita.shared.exceptions;
 
 @SuppressWarnings("serial")
 public class ForbiddenException extends RuntimeException {
