+++ src/main/java/com/project/miinhareceita/shared/exceptions/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.project.miinhareceita.services.exceptions;
+package com.project.miinhareceita.shared.exceptions;
 
 @SuppressWarnings("serial")
 public class ResourceNotFoundException extends RuntimeException {
